Remove deps on `node-sass`, use `sass` instead. (#20808)
`node-sass` is deprecated and sort-of depends on being able to download
precompiled binaries that match the version of node in use. `sass` is
the replacement.

The update to calypso-build 9.0.0 in #20789 took care of most of the
dependencies on `node-sass` in the monorepo. This cleans up the rest.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1167056866