Use Node 16.7.0 in tooling (#20827)
To match the version of Node used in Calypso.

This shouldn't change the behavior of any of the code itself.
The main "risk" as far as incompatibility is if future PRs to
eslint-changed start using Node 16 features. Everything else that's
published is targeted at browser environments or is an action where our
Node 14 was already past GitHub's Node 12.

Additional changes:

* Replace jsdom-global with global-jsdom.
  It seems the former is no longer being updated.
* Hack around a React bug.
* Move Jetpack-specific test init out of the runner and into Jetpack.
* cli docker: Create `.env` with `openSync`.
  `createWriteStream` is async, and it wasn't creating the file in time for
  the following docker command to see it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1175160067