storybook: Test that deps match projects.js (#24188)
We have a comment at the top of projects.js asking that people keep the
list in sync with the list of dependencies in composer.json, but it
seems this is easy to miss. Let's also add a test checking for the
match.

For example, #23780 missed updating composer.json when adding to the
list, and since the dependency was missing #24078 didn't know to try
rebuilding storybook when plugins/protect was changed, which allowed it
to be merged and break the build in master (necessitating #24160).

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2263957138