cli: Detect if run from another checkout (#24950)
The CLI "link" feature is useful if you only use one checkout, but it
can cause confusion if you ever try to use a second checkout.

We can DWIM by having the CLI check if the cwd is within another
monorepo checkout, and if so run that checkout's copy of the CLI instead
of the linked copy. This winds up having to be done in a halfway-hacky
manner since node doesn't expose C's `exec()` family of functions, but
it should work.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2617162269