renovate: Use configured version of PHP when running composer (#25112)
In the absence of configuration or certain dependencies, renovate
currently defaults to running composer with PHP 7.4. Or, more
specifically, it defaults to whatever is the 'latest' tag on the
`renovate/php` docker package.

We do everything else with PHP 8.0, so configure renovate to use that
too.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2741191776