Gather support references: add option to send a Slack message (#25388)
* Gather support references: add option to send a Slack message

When issuees start gathering a lot of tickets (at least 10), let's send a Slack message so the issue can be escalated if needed.

We only want to send that Slack message so after doing it, we add a note to the comment that lists all support references to indicate it was escalated. Later, we search for that note before we try to send a new one.
Internal reference: pciE2j-1iy-p2

* Return early when a note has already been sent

We need to note that a message was sent, without sending a new one.

* Update escalation message and extract it

See https://github.com/Automattic/jetpack/pull/25388/files#r938762188

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2804686657