pnpmfile: Don't do "webpack without webpack-cli" check with pnpx (#25807)
When running `pnpx`, pnpm creates a temporary directory to install the
requested program and its deps to run it, but it still applies our
`.pnpmfile.cjs`. The "webpack without webpack-cli" check doesn't make
sense in that context, so detect that and don't run it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2921486051