phpcs: Re-enable WordPress.DB.DirectDatabaseQuery sniff (#25806)
We disabled this during the big phpcs cleanup project because we didn't
need to or want to fix it at the time. That project is done now, and the
rule is thought to be useful enough to not permanently disable, so we
should re-enable it so new instances don't get added.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2934977999