tools: Be more explicit about desired versions in version-packages.sh (#25989)
Currently `tools/version-packages.sh` just updates to whatever is the
latest on Packagist, meaning that if we've just pushed new versions
(which is usually the case) then we have to manually make sure Packagist
has gotten the updates. If it hasn't, we have an existing check that
will handle a missing minor or major bump, but not a patch bump which
may still contain import bug fixes that we want to include in the
release.

We already know which versions we expect to see, since they're listed at
the top of each package's changelog. So let's extract them, and
explicitly request them of Composer so it will error out if the expected
version isn't available from Packagist.

Then let's also check that any indirect deps got updated correctly in
the lockfile too.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/2973756884