ci: Check plugin monorepo dep versions (#26117)
In Jetpack 11.3-beta3 we accidentally had some packages locked to
"dev-trunk" instead of the expected release versions. This adds a CI
check to try to catch when that happens.

For pushes to trunk and PRs for merge into trunk, all monorepo packages
are expected to be locked to dev-trunk in all plugins.

For pushes to a release branch (and any PRs against release branches,
should there be any), all monorepo packages for the plugin associated
with that branch are expected to be locked to a non-dev version that
compares greater than or equal to the latest version in the package's
changelog.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3025022729