build: Post wpcom testing reminder on all Jetpack PRs (#26194)
When this was added in #25310, it was set up to only post on PRs that
touched de-Fusioned files, with the intention that we'd update it once
de-Fusioning was well under way. That time is now.

Also print status output always, not just when debugging is enabled.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3047561614