Social: Change the connections endpoint to connection-test-results (#26274)
* Social: Change the connections endpoint to connection-test-results

After merging #26216, we discovered that Jetpack Social was not using
the same API endpoint to get the connection data in the editor. The one
used by Jetpack and WPCOM already provided information about the health
of the connection.

This change reverts the addition of the `is_healthy` property in the
`/publicize/connections` endpoint and updates Jetpack Social to use the
`/publicize/connection-test-results` endpoint.

* Verion bump

* Update the composer.lock

Co-authored-by: sdixon194 <steve.dixon@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3107399415