VideoPress: Fix video counter on VideoPress library page (#26365)
* Support videosOnPage argument on VideoLibraryWrapper component

* Add optional totalVideos prop on VideoLibrary components

* Use new totalVideos prop on the VideoPressLibrary component

The VideoPressLibrary component now expects the new totalVideos prop, and uses it while rendering the VideoLibraryWrapper component.

It also uses the `videos` array length as the number of videos available on the current page.

* Set totalVideos prop on the Video Library components

* Use new totalVideos prop on the LocalLibrary component

* Remove unnecessary videosOnPage prop

The videosOnPage counter is not necessary, based on the Figma designs for this page.

* Set proper plural and i18n on the video counter

* Include changelog file

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3108750552