Social: Require a user connection to use the plugin. (#26543)
* Social: Require a user connection to use the plugin.

If only a site connection is present (for example, when Jetpack Boost
has been set up), then it was causing errors as the editor was loaded.
This was down to us trying to retrieve and test the status of, social
media connections for a non-existent user.

This change adds some extra logic to ensure that we have a user
connection present before we load the editor extension or the main admin
page.

* Without a user connection, don't load Publicize.

This stops Publicize from being initialised in the Classic editor
without a user connection.

* changelog

* Fix the test by mocking the user connection.

* Add a test to check Publicize is not configured without a connection.

* [not verified] Only render required state when there is no connection.

Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3183479695