Social: Add the showNudge reducer (#26635)
* Social: Add the showNudge reducer

The reducer for showNudge was missing, so it would default back to
`false` if a value in the store changed. This meant that after clicking
'Start for Free' on the pricing page, it would seem like the user had a
paid plan.

* Move showNudge to jetpackSettings in the store.

* Remove the default.

* Add a default in case showNudge isn't present for any reason.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3189735137