Client: Update Section Header component to Typescript file (#26644)
* [not verified] Client: Update Section Header component to Typescript file

Why: To reap the benefits of TypeScript (such as added safety and automation) for this file.
Also, a good way to practice creating a PR for Jetpack.

* [not verified] Set typed label property as required in SectionHeaderProps interface

The label property is passed everywhere this component is implemented, which makes sense given the component displays the value of the label property as the SectionHeader content

Co-authored-by: Konstantin Obenland <obenland@gmx.de>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3228842573