tool-setup: Configure composer to not use GitHub's API (retry) (#26788)
We had to revert #26763 because setting use-github-api caused a change
to the root composer.lock file, apparently not using GH's API also makes
it lock to a slightly different URL.

This reinstates the change, with the root composer.json also explicitly
setting use-github-api so composer in people's local environments will
behave the same way.

Should a project someday both use a "vcs" repository and check in
composer.lock, it would need to do the same thing. As it is, none of the
plugins use a vcs repository entry and none of the packages using a vcs
repository entry check in composer.lock, so they're all good.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3235519956