Bugfix: font sizes in payment-buttons taking precedence over contained buttons (#26839)
The payment-buttons typography feature introduced in #26580 had a bug
which meant that the font size in the payment-buttons block was
overriding font sizes set on the buttons when a preset font size was
being used. This bug occured when viewing the site, but not when
editing.

The bug was caused by the view CSS selector not being quite specific
enough.

The font size of the button blocks when using a preset works by adding a
has-{presetName}-font-size class to the element in question, where
presetName varies depending upon the theme. To fix this bug whenever a
preset font size is used a class called has-custom-font-size is added in
both JS and PHP. This gives a known stable class name which is targeted
with CSS.

Fixes #26748

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3263833860