Jetpack: Sync site endpoints with wpcom (#27059)
Not sure if we really need these in Jetpack, as the get-site endpoint
was broken since April 2021 (#19545) and no one noticed. But we may as
well sync them up for de-Fusioning since it's not that hard to do, the
removal can happen later.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3324197947