Publicize: Fix the is_healthy logic in the classic editor. (#27159)
We removed the is_healthy property from the API and changed the logic
when rendering the form in the classic editor, but it was wrong. This
was only discovered as the API was updated when the Jetpack Beta was
deployed to WPCOM.

This PR fixes the logic.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3348076392