actions: Fix slack-workflow-failed workflow (again) (#27234)
In #26800 we added use of a shared function to replace the deprecated
set-output command. Then in #27205 we tried to fix it, but that just
revealed another problem.

Hopefully this is the last problem. It'd be nice if this weren't so hard
to test.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3386327165