VideoPress: Expose video visibility flag on VideoPress video data (#27349)
* Add Data method to return the value for the VideoPress site privacy setting

* Expose video_is_private field on the VideoPress data for every video

This field takes into account the value of the video privacy setting, as well as the VideoPress site-wide privacy setting.

The site-wide setting takes precedence over the video one, so when the site setting is set to private all the videos become private, regardless of the individual settings of the videos.

* Add changelog file

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3439810153