tools: Fix error exit in create-release-branch.sh (#27771)
When we run `git log -p`, it probably uses `less` to page the output.
In some cases when you quit `less` it will send a SIGPIPE back to the
git process, causing it to exit with code 141, which then terminates the
script due to the `-eo pipefail`.

The fix is simple enough, append `|| true` so the command as a whole
always ends with a 0 status so the script can continue.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3630601730