Jetpack: Remove require of missing file in tests/php/bootstrap.php (#27782)
PR #27407 missed removing a require of wpcom-helper-functions.php from
tests/php/bootstrap.php. Remove it.

It's odd that this didn't make tests fail. It turns out that's a bug in
the Patchwork library, it prints a warning but then treats it as an
empty file.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3636856121