tests: Only update dev deps when preparing plugins (#27818)
Our composer.lock files are currently generated with PHP 8.0, but we run
tests with earlier versions too. To make this work, we run `composer update`
if `composer check-platform-reqs` indicates that the install is going to
fail.

But that causes a problem when running the tests on release branches,
where newer versions of the monorepo packages might exist for it to
incorrectly update to. What we really want to do is only update the
dev-dependencies, since if any of the non-dev deps don't work with a PHP
version that probably means the plugin is broken anyway.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3680206252