Fix Goodreads' custom widget not finding target element (#28946)
* Fix Goodreads custom widget not appended properly to HTML element #8071

This is necessary, since the Goodreads Custom Widget script
replaces dashes with underscores in the widget_id parameter.
This causes the custom widget script to not find the target already
present in the DOM, and appends the widget markup
right after the script.

* changelog

* Fix comment typo

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4178868599