tools: Strengthen report about old PHP in dev env check (#29003)
The note that "Version 8.0 or later is recommended" issued by
`check-development-environment.sh` when using PHP 7.4 is being read to
imply that 7.4 should work, which is not the case for our tooling (it
just should be ok for running plugins).

The message will now state "Version 8.0 or later is recommended. Tooling
may raise errors with earlier versions."

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4201110351