tools: Allow selecting more than one but not all plugins in create-release-branch.sh (#29058)
Turns out this is something we're going to be doing a lot, releasing
Jetpack and mu-wpcom-plugin but not Backup or Social. So now we can
either list multiple plugins on the command line or be prompted for each
candidate plugin if only one is supplied.

Also, update the bash version check to require 4.3 (released 9 years
ago) instead of just 4.0, as this uses "unset" which was introduced in
that version.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4232538690