VideoPress: support "private" as video "site default" privacy on public Atomic sites (#29104)
* Create function to determine if VideoPress videos are private by default

This is the start of the refactoring to improve how we handle privacy on Atomic sites.

* Add check for Atomic sites

If it's a private atomic site, the site default privacy setting is private as well. If it's public, treats like a Jetpack site and look for the `videopress_private_enabled_for_site` option.

* Include check for Simple sites

For Simple sites, the site privacy controls the site default video privacy. If the site is private, the default privacy for videos is private as well. The other way around is also true.

* Use the same logic from the Ajax class to determine video privacy default

* Rely on the Data class to determine site default video privacy

* Bump versions

* Fix the return value for private atomic sites

If the site is Atomic and is private, the function should return true signaling that the videos are private for the site.

* Force return value to bool type

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4297913038