tools: changelogger-release should ask if there are no changes (#29425)
When running `changelogger-release.sh` on a plugin, it may be that the
plugin itself has no changes but packages it depends on do so we still
want to do a release.

So, in the case that `changelogger-release.sh` finds nothing to do for
its command line argument, have it ask the user whether they want to do
a release anyway. If so, generate a change entry for "Internal updates"
to be used in the changelog.

Also, instead of copy-pasing the same code for figuring out which
`--type` to pass to changelogger for the 4th time, let's make a function
to handle it and use that everywhere.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4388107312