Fix Jetpack tests (#29695)
PR #29527 accidentally broke Jetpack's tests by adding a call to
`Patchwork\restoreAll()`, which removes our redefinitions of `exit()`
and `die()` that other tests are relying on.

The immediate fix for that is to only restore the redefinition that was
defined in that test.

Then, to prevent this from happening again, we use
`->addExpirationHandler()` to catch when our redefinitions get removed
and exit immediately with an unsuccessful code.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4513187488