gardening; Don't try to clone PR branch after close (#30056)
In #21213 we updated the action to not try to clone the PR on the
`closed` event itself, but the `pull_request_target` event can also fire
post-close for `labeled` and `edited` and maybe others.

So instead of checking the action type, look at the PR's state field to
decide whether to clone or not.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4679322180