Release Script: Support multiple plugins (#29323)
This updates the release-plugin.sh script to add support for multiple plugins, and will automatically add the mu-wpcom-plugin if we're releasing jetpack.

When prompting for versions in `changelogger-release.sh`, pre-fill with the current version since
that's most commonly what's going to be used. Also don't bother
confirming when we're using the current version, that's the usual case.

---------

Co-authored-by: Brad Jorsch <anomiex@users.noreply.github.com>
Co-authored-by: Brandon Kraft <public@brandonkraft.com>
Co-authored-by: Brad Jorsch <brad.jorsch@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4682626810