actions: Improve build and post-build workflows on forks (#30232)
The "Trigger TeamCity build" cannot run on a fork as it needs secrets,
so we'll have to skip it.

The post-build workflow is fine to run, but it will think there are no
plugins because the "Create artifact for Jetpack Beta plugin" job
doesn't run and it looks for that job's artifact to decide if there were
plugins. It doesn't actually _use_ it, it just looks for its presence.
While we could download and read the zip file instead, it seems easier
to just create a dummy artifact to serve as the signal instead.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4759365087