Related Posts: Fix context for use with blocks (#30682)
Adds a second "block context" for use in block rendering. The existing context is still generated for use by the non-block version of the output. With the new context, we also properly set up a link to the category, tag, or comments as appropriate.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5056349125