build: Short circuit Composer 304s (#30824)
We're getting close to 100 projects in the monorepo, 84 of which don't
have a composer.lock. For each one of those a part of the build is to
run `composer update`, which means Composer hits packagist for every
dependency's metadata to check for possible new versions, just to get
back a 304 response because usually there aren't any in the few minutes
since it last checked. The RTTs to get those 304s can add up to a few
minutes across all those builds.

In this PR we add a simple JS transparent proxy in CI that will remember
URLs that have been queried during the run and respond with an immediate
304 without hitting the network again, and a composer plugin that
rewrites the metadata request URLs to fetch from the proxy.

In a local test out of 12847 metadata requests during the course of a
full build only 336 had to be sent to packagist, the other 12511 were
repeats that were short-circuited. Eight of the metadata files were
requested 196 times each!

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5071222956