ci: Remove required-reviews CI job (#31118)
This was made un-required months ago, and despite the list being trimmed
down was never re-required. There's no point in continuing to spend
resources on running it if it's not going to be checked, so let's remove
it.

Ref: p1684435201776199/1684417733.819429-slack-CBG1CP4EN

We'll keep the actual action in the monorepo though. It seems to be
something that's actually attracting third-party contributions and
doesn't require much maintenance.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5139256775