Update monorepo to Composer 2.5 and PHP 8.2 (#31150)
Composer 2.5 has been out for a while now, and doesn't even have an API
version update in the lockfile.

PHP 8.0 security support ends in November, and we've been running CI
tests with 8.1 and 8.2 for a while. Let's update that too.

We still need to run the WP previous check with PHP 8.0, though, as 6.1
wasn't compatible with later versions.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5325071700