Social Review Prompt: Ensure it is still shown with Jetpack Active (#31456)
* Social Review Prompt: Ensure it is still shown with Jetpack Active

We had to revert #29910 in #30101 because the review prompt state was
stamping on the rest of the initial state object that's used by
Social/Publicize.

This change adds the Post Publish panel with the review prompt to the
publicize-components package and merges the required state in the Social
plugin, so that if both plugins are active the review prompt is still
shown in Jetpack.

* Fix the package versions

* Add the @wordpress/edit-post dependency

* Fix package versions

* Fix package version

* Update the script handle

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5338246295