tools: Add checks for filename collisions on case-insensitive systems (#31622)
In the pre-push hook, we check the commit being pushed for collisions,
as checking out the commit for review will confuse git on Mac OS and
other case-insensitive systems.

In the mirror build, we check the output as collisions will likely cause
problems if the package/plugin/whatever is used on a case-insensitive
system.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5413108658