renovate: Remove "close old PRs" workflow (#31885)
We added this in #19307 because renovate had a bug where it would lose
track of old PRs once they were beyond the latest 1000 in the repo.
Supposedly this is fixed now, so no need for the cleanup workflow
anymore.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5552825220