CRM: Add hash router to Automations admin (#31886)
* Add hash router

* Add home redirect

* Add changelog

* Make the Admin page test a tad more specific

* Move React Router to when we render the app itself

The main reason we do this is to we can wrap our tests with <MemoryRouter> to gain more control in Jest.

Resource: https://v5.reactrouter.com/web/guides/testing (it's v5 and not v6, but the same principles seems to apply)

* changelog

* Delete duplicate automations admin-page-init.php file

This was a leftover from a pretty chaotic merge conflict

* Remove prettier change in other packages/projects after merge conflict

* changelog

* Remove prettier change in other packages/projects after merge conflict

Second attempt: running this with --no-verify to avoid the same thing happening again.

* Trying to match external "non-prettified" formatting

Technically, third attempt.

* Remove Forms and Jetpack changelog since we no longer include changes

---------

Co-authored-by: André Kallehauge <3846700+kallehauge@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5603645169