actions: Close stale PRs after an additional month of no activity (#32113)
If no one cares enough about the PR to at least to a trunk merge, we may
as well close it. Chances are we'd have been doing about the same
manually anyway.

The message posted when marking a PR as stale is adjusted to explicitly
request the trunk merge, and to warn that automatic closure will happen
in a month.

When closing, a message is posted inviting people to restore the branch
and reopen the PR if they want to resume work on it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5684388396