Jetpack CRM: Use WP_UnitTestCase instead of WorDBless (#32498)
* Proof of concept of using WP_UnitTestCase instead of WorDBless

* changelog

* Remove WorDBless package dependency

We also add 81 polyfill since to avoid "enum_exists() do not exist" errors that was introduced with the swap.

* Use new base class in tmp test class

* Re-add REST tests

The recently added polyfill made these existing tests work.

* Replace WorDBless\BaseTestCase with JPCRM_Base_Test_Case

* Check if table exist before we reset/nuke

"zbs_global_total_trans" is created in a paid extension but we assume that it exists when running our core logic which will
provide an error.
The real fix would be to move it add listen to new do_action() hooks, but that's outside of the scope for this PR.

* Do not use prepare for TRUNCATE and DROP

We can't use "$wpdb->prepare()" because it will add single quotes around the table name which breaks the query.

* Restore original ZBS global value for each test

This is a requirement in hopes of having just some backwards compatibility with existing tests.
The demo REST API tests have also been updated to no longer use mocks, so we have real tests that create and fetch contacts.

* Delete tmp test and enable all existing tests

* Separate base integration test logic into its own class

We do this to 2-3x the speed of which all of our tests complete since we do not have to clear all tables unnecessarily after
each test.

* Revert zeroBSCRM_database_nuke update

We did not end up using zeroBSCRM_database_nuke(), so it introduces noice to the PR.

* Make test root logic a 1:1 copy of the Jetpack plugin

* Remove symfony/polyfill-php81 dev dependency

* Update Transaction_Condition_Test to extend new test case

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5883239518