Fix a Global Styles UI performance issue (#32641)
* Fix a Global Styles UI performance issue

Instead of loading the Jetpack blocks CSS for the Site Editor using the admin_init hook, we'll switch to the `enqueue_block_editor_assets` hook. In this manner, the CSS script is loaded in the iframe, instead of the main site-editor.php page.

By doing so, the UI for Global Styles is more responsive.

* Remove workaround and use `enqueue_block_assets` directly.

* Add is_admin guard and add docblocks explaining why we need this guard and what's the difference between the enqueue_block_assets and enqueue_block_editor_assets hooks.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5977164315