ci: On release branches, build/test relevant plugins only (#32775)
When we're running CI builds and tests on pushes to release branches, we
only really care about the stuff that's being released on that branch.
So only consider those projects as "changed" instead of building or
testing everything in the monorepo.

Also, for tests, consider the dependencies of release branch stuff as
changed since any cherry-picks that wind up in the release should also
be tested.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6096905806