ci: Allow specifying events for turnstile (#33189)
Currently our turnstile action as used for the Gardening workflow groups
issues, issue comments, and pushes to trunk together, which can result
in trunk pushes timing out when a big batch of issues are filed.

The solution here is to allow specifying a list of events to process
and/or ignore to turnstile, so when we're processing an issue or issue
comment we can tell it to only consider other workflows for those events
and when processing a trunk push we can tell it to ignore those events.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6303798564