ci: Work around a node 18.18.0 bug (#33353)
It appears that node 18.18.0 has done something that breaks
`thread-loader`, causing webpack to hang. It can be worked around for
now by setting an environment variable, so let's do that.

In some quick testing the bug is also in node 20.3.0 to 20.6.1, but is
fixed in 20.7.0.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/6330246114