ci: Fix monorepo auto-tagger (#34439)
GitHub has a limit on the number of tags that can be pushed in one
command. While we could change the setting, it's probably better to work
with it in the workflow and batch the tag pushes.

Let's also alert in Slack if the auto-tagging workflow fails.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7103863777