ci: Handle concurrency and improve performance in Monorepo Autotagger workflow (#35229)
The `git push` of the new tags is taking a surprisingly long time, long
enough that e.g. the trunk merge into `prerelease` before creating the
PR for it can wind up happening before the previous commit's run
finishes.

Let's use turnstile to avoid the occasional failure in the tag push. But
to avoid unnecessary waiting, split the existing code into separate
"determine which tags are needed" and "tag the tags" steps with the
turnstile in between.

But as for performace, it seems the treeless clone is what's making it
be so slow. For some reason on push it's fetching all the trees. Let's
switch to blobless clone instead, which is slightly slower initially but
doesn't seem to have that problem.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7657332039