Jetpack: Fix woocommerce-analytics script enqueuing (#35471)
In #34072 we replaced a call to the being-deprecated
`Assets::enqueue_async_script()` method, which is ok with being passed a
URL, with `Assets::register_script()`, which requires a pathname. This
results in a PHP warning about filemtime and an incorrect script being
enqueued.

Switch it to calling WP Core's `wp_enqueue_script()` instead. And name
the variable correctly for good measure.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7803650232