Have tsc-built packages use moduleResolution:nodenext (#35453)
For the built JS and declarations files to be usable in other TypeScript
projects that aren't using moduleResolution:bundler, we need to build
them with the stricter moduleResolution:nodenext.

This means that we can no longer do extensionless imports or directory
imports, everything has to point to a specific file.

But not necessarily the actual file on the filesystem, as tsc requires
files be referred to like "foo.js" even if the actual file is "foo.ts"
or "foo.tsx". Webpack then needs reconfiguring to do the same thing.

A followup will create a new preset in `tools/js-tools/` and make some
other tooling cleanups, but first I wanted to get these modules fixed.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7906251783