webpack-config: Add fork-ts-checker-webpack-plugin option (#35476)
We've decided (pdWQjU-FO-p2) to prefer `@babel/preset-typescript` over
`ts-loader` for dealing with TypeScript in the context of Webpack. For
generating the definition files, we're generally going to use
`fork-ts-checker-webpack-plugin` (although `tsc --emitDeclarationOnly`
is also an option).

To make that as easy as possible for people to work with, let's make it
an option in the shared webpack config.

Also, since the code was starting to get cluttered, I sorted all the
"plugins" into alphabetical order to make it easier to find.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7919211088