Sharing block: avoid error when a service is not available (#36077)
This is a check that likely will not be that useful to site owners, since we will be adding services, not necessarily removing them.
However, during development and switching between branches, it can happen that a service you've added to the sharing block is not available anymore.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8205246409