stub-generator: Handle functions calling `func_get_args()` (#36870)
If a function or method calls `func_get_args()`, Phan treats it as if it
has an implicit varargs argument at the end of its parameter list. But
when we stub the function it can no longer see the `func_get_args()`
call so it tries to incorrect force the declared parameter list.

So let's do the same thing when generating the stubs, if there's a
`func_get_args()` call (and no varargs param already) let's add one.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8658899618