Create automattic/patchwork-redefine-exit package (#37476)
This is an extraction and cleanup of the code we've long used in
Jetpack-the-plugin to avoid dying in integration-ish tests when
WordPress code calls `exit` or `die`.

Now that we're splitting more code out of Jetpack, we're starting to
need this in packages too.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/9193387973

Upstream-Ref: Automattic/jetpack@213ee958066a4cf6d17feb6c8d550347c1e1f53e