tools: Default some prompts to "yes" (#37603)
Some of the prompts, such as the one about pushing the release branch
after creating it, really should default to "yes" so if someone is just
hitting enter they do the thing that's right most of the time.

So, add an option to the `proceed_p` shell function to specify that the
default should be "yes", and use that in various places where it seems
to make sense.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/9566763820

Upstream-Ref: Automattic/jetpack@81009c4dcc7382b583a56699e48a612f9a1e3a51