tools: Improve performance of `check-intra-monorepo-deps.sh -u` (#38069)
The redefinition of `init_changelogger` wasn't taking effect for the
early steps because `changelogger` was being run in a subshell. Add an
`init_changelogger` call before that so it can take effect.

Waiting for the composer.lock update is a large portion of the script's
time. Fork these updates into the background so they can run in parallel
with each other and with the rest of the script.

On my laptop, this reduces the time taken for a simple run from ~36.5s
to ~7.6s.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/9695718323

Upstream-Ref: Automattic/jetpack@84f7600a33dba6329bf886c440157ff31aba3bfc