tools: Force releases of js-package dependents (#41572)
Our js-packages are usually bundled when included in (php-)packages and
plugins, rather than being used directly at runtime. This can lead to an
edge case in releases, if a js-package has been updated and the plugin
being released depends on that, but the php-package that bundles the
js-package has no changes and so does not get released.

To avoid this, when a js-package is being prepared for release, any of
its dependents should also be released. We go about that using a
solution similar to that from #33665, which fixed a similar edge case.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/13166671565

Upstream-Ref: Automattic/jetpack@98233475eeefb13d74c6432ffccb5eb8e19ce529