build: Deepen checkout for js-packages/social-logos/src/svg (#42394)
The font for js-packages/social-logos embeds a timestamp. To avoid
unnecessarily changing the file, we derive the timestamp from the latest
commit to the `src/svg` directory.

Starting with 043235bd, every commit has been changing the file anyway.
It turns out this is because the existing deepening of the checkout for
packages' -alpha versions is typically doing 2000 commits, and that's
the 2000th commit since the last commit that touched social-logos's
`src/svg` dir. So the git timestamp check was returning the timestamp of
the graft point rather than 7d77b839 like it's supposed to.

The solution is to make sure the checkout is deepened far enough to
reach an actual commit to that `src/svg`.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/13861159576

Upstream-Ref: Automattic/jetpack@2e253f4a5eaefa8b0cd605cc8822bc70e09ccfa2