cli: Update `jetpack docker phpunit` command (#43437)
The command now takes a "target" argument specifying which tests to run.
The separate commands "phpunit-jp-multisite", "phpunit-crm", and so on
are deprecated.

This also updates the docs and messages from the CLI to note that
`jetpack test` should be used to run tests for anything not supported by
`jetpack docker phpunit`, with instructions added in
`docs/development-environment.md` for how to use `jetpack docker sh` to
run them inside the container if you really want to.

Unfortunately the help output from yargs (at least the version we
currently use) is fairly lacking. But that can be looked at in a
separate PR.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/15003430541

Upstream-Ref: Automattic/jetpack@2985a9289dd78cf3161405c3a1fa827320872166