sass: Fix "slash as division" deprecation warning (#43624)
CSS uses `/` as a separator in some contexts. Sass treats it as division
when it's not a separator, but it can be confusing which is the case, so
sass v2 is removing slash-as-division in favor of the `math.div()`
function.

https://sass-lang.com/documentation/breaking-changes/slash-div/

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/15260149552

Upstream-Ref: Automattic/jetpack@73962537557dd630d0372e53dcbd7b2779c5f21a