fix: Private site returns valid access cookie expiration (#44328)
* fix: Private site returns valid access cookie expiration

The `set_logged_in_cookie` hook signature includes a second "expires"
parameter as an integer, not an array. It is unclear why this originally
used an array as a second parameter and if/why it worked previously.

From discussing with the author, this was likely an erroneous change in:

https://github.com/Automattic/wpcomsh/commit/cabdb38b1474db7381d80a7ac1c3357960635331#diff-c26ac91049f9950230fd95e1577719794c983a4ae493e2875eadc853378a200aR429-R431

* test: Assert private site logic

* feat: Private site access cookie uses `expiration` rather than `expires`

By using `expiration`, we can leverage the 12-hour grace period included
the `expires` value provides. I.e., when `expiration` lapses, we could
request a new cookie during the grace period provided by `expires`,
where the cookie remains valid.

* changelog

* test: Correct cookie expiration test logic

The expiration value is intended to be an offset--a duration from the
current time, not a specific time.

https://developer.wordpress.org/reference/hooks/auth_cookie_expiration/

* style: Address misaligned documentation

* test: Mitigate test instability by assert delta instead of absolute time

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/16345871294

Upstream-Ref: Automattic/jetpack@64a1fdc3303b1abf0ed07357ec241748c15740c4