cli: Fix `jetpack phan --include-analysis-file-list` (#46203)
If files in projects and not in projects are included in the list, it
was getting confused about which projects it's supposed to run on. Fix
that.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/19944211427

Upstream-Ref: Automattic/jetpack@f1968a9bb6fd03908dcfd548677ecdc63280c5b5