Remove unneeded hoisting and monorepo-root deps (#46377)
These were added in #46206 in an attempt to fix an experimental build
using `wp-build`.

Any dependencies in the monorepo root, and anything hoisted by
`publicHoistPattern`, are going to be found by any project in the
monorepo, which is very likely to lead to people forgetting to declare
needed dependencies in their projects. Which is why we very carefully
minimize the dependencies in the monorepo root in the first place.

If `wp-build` seems to require those things, either it's broken with
pnpm (and is also probably broken with yarn P'n'P) or you have it
configured incorrectly.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/20380686024

Upstream-Ref: Automattic/jetpack@652f4bbfb76506de1b519ac41a1899d77c084b15